requirements.txt
torch
torchvision
numpy
matplotlib
# GUI toolkit, so matplotlib can display graphics
PyQt5==5.9.2
pyyaml
